requirement.txt
mysql.connector
python-dotenv